+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+ - Add support for empty framebuffers with the `EmptyFrameBuffer` type.
  - Fix an OpenGL error when mapping a buffer with `map_read`.
 
 ## Version 0.8.7 (2015-08-27)